Add: BirdReactionGraphInfo ViewModel 추가
