+++ AKCOO/Domain/ViewModel/Bird/BirdReactionGraphInfo.swift
@@ -1,0 +1,15 @@
+//
+//  BirdReactionGraphInfo.swift
+//  AKCOO
+//
+//  Created by Anjin on 11/26/24.
+//
+
+import Foundation
+
+struct BirdReactionGraphInfo {
+  let criteriaTitle: String
+  let minimum: Double?
+  let maximum: Double?
+  let userAmount: Double
+}
